Update patch set 2
Patch Set 2:

When app sending sequence in on/off pulses it looks like this: 40,15,40,15,1300 and in fractions it looks like: 1000,375,1000,375,32500. So if value is big enough >255 and not last then sequence most likely in fractions otherwise in on/off pulses. Last one is ending one which pretty big. Maybe comparing values should be tunned a bit. But for now I have working smartremote and smartir. Which uses different types of sequences.

Patch-set: 2