Update patch set 2
Patch Set 2: Code-Review-1

Can I rephrase?

You are trying to guess in the HAL whether or not the user-space application is sending a sequence in terms of pulses or terms of ms?

If so, the API specifies that it should be sent in units of "ms".

Any application that isn't following that specification should be encouraged to be fixed rather than adding code to try to force the incorrect app to work.

Patch-set: 2
Reviewer: Christopher R. Palmer <6351@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Code-Review=-1