Update patch set 2
Patch Set 2:

Maybe the ones that are working now don't work on stock and the ones that don't work now do work on stock?

SmartIR doesn't work with the previous code but they have the ability to tag a customer's account with "legacy" and then the application itself switches to using pulses.

Patch-set: 2