Update patch set 1
Patch Set 1:

On mondrianwifi, this does not load my widevine.mdt.

To test, I am using Google Play Movies & TV and attempting to download a video.  The log does show V12 DRM being used but doesn't work and has these errors in the log:

09-20 15:33:34.806 E/QSEECOMAPI: (  246): Error::Cannot open the file /vendor/firmware/widevine.mdt

09-20 15:33:34.806 E/QSEECOMAPI: (  246): Error::Cannot open the file /system/etc/firmware/widevine.mdt

If this patch is working for everyone else then I'll leave my mondrianwifi specific commit in my device repo and this commit is fine to submit as it won't hurt me, just doesn't help me with widevine.

Is it really loading widevine.mdt for you?  For me, that only gets loaded when used, so there is no error either way on boot.

Patch-set: 1