Update patch set 1
Patch Set 1:

Exporting the header is unlikely to change anything (all it does it copy the header to the $OUT directory).

Unless I completely missed something, this change should also be very unlike to change anything (it builds a shared library that no one is using).

How sure are you that it is really the cause of your problem?

Patch-set: 1