consumerir: Dynamically grow an appropriately sized buffer
The IR pulse sequences are written to a local buffer to be
passed to the kernel.

Previously, we doubled the size of the buffer to work around a
crash that was occurring for long sequences.

Another crash has been found when using Smart IR Remote and the
POWER function of the TX P42X10E remote by Panasonic (TV category).

Rather than double the buffer again and wait for another counter-
example, dynamically allocate the buffer and grow it as needed
to handle any IR sequence.

Change-Id: Id4fb799d8a5ad4242e04a68e6cc6436a14f4d2d0