Update patch set 3
Patch Set 3:

Yes, I pushed the revert a few minutes ago.

FYI, Ethan, the problem was that I actually require that some of the firmware be in the new location and I created a symlink to hack around the problem until we figured out what to do with it.

Patch-set: 3