Update patch set 6
Patch Set 6:

Ethan, given that there are no #ifdefs in the code, this seems like it really is msm8974-common at this point in time.

AFAIK, all the devices need all the changes in here (hlte and mondrianwifi definitely need all the fixes in here, that's why we merged it).

Patch-set: 6