Update patch set 6
Patch Set 6:

The header being located elsewhere doesn't bug me *EXCEPT* for the fact that the wrapper is using things that must be defined in the header.  That does seem very wrong.

I considered suggesting CameraParametersExtraTarget.h which defines the target specific parameters (and therefore CameraParametersExtra can remain with the wrapper and it builds the macros based on what it needs locally and what the ExtraTarget defines).

But that sounded like over-kill.

I'm also not against just pushing this up one level to the devices, if that seems cleaner.

Patch-set: 6