Update patch set 6
Patch Set 6:

At the birth of this repo it was for hlte and mondrianwifi only.  We were still inheriting from msm8960-common and having to override several items.  Chris and I put this repo together, and at that time it made sense to push anything common here.
Now that we are supporting more devices this may not make sense anymore.

Patch-set: 6