Update patch set 6
Patch Set 6: Code-Review+1

I think the way we are going to do it is fine. its how they are doing it with the JF. leave the wrapper in msm8974-common or now. move the header...if it becomes an issue later we can always move the wrapper down. but less work on everyone the better in my point. except for a few different symbols here and there that need to be defined in header the camera module seems pretty generic at this point. No ifdefs or anything like that.

Patch-set: 6