Update patch set 5
Patch Set 5: Code-Review+1

Tony, I pulled out the build.prop change which was working to try to get video recording fixed and put it into a separate commit to work on fixing the video recording.

With that done, I think this is ready to be merged as simply a codec update and not an attempt to fix the camera.

Patch-set: 5
Label: Code-Review=+1