Update patch set 1
Patch Set 1:

For example:

https://play.google.com/store/apps/details?id=com.madfingergames.unkilled&hl=en

when played with a2dp output it sounds horrendous.  I haven't debugged this much but I had the same problem on the zenfone2 on M with most games.  On that device it seemed like the buffer sizes reported by the HAL (vendor blob HAL with no source code available) were larger than what was required to work.

On that device I had the same problem on primary and a2dp output.  I wrote a wrapper hal for primary.*.so and hacked it to return buffer_size / 4 when opened with flag_fast and that worked to fix everything other than a2dp.  I didn't bother to try the same thing for a2dp to see if it would and instead just removed the permission.

I suspect it is something similar here but I'm not sure I really care enough to spend the time debugging it.

Patch-set: 1