Update patch set 1
Patch Set 1: Code-Review-1

This builds fine for me on mondrianwifi.

Maybe you just need to add the include in your device_samsung_klte/init/init_klte.cpp since that's where it looks like it's really needed anyway?

On a different note...

Generally, I'm not sure that this code should be here at all.  Ethan specifically requested that this repo be a "light, config only" repo and we have moved all the code out of here and up to devices.

Michael Bestas added this without much review or notice and is the only piece of code currently in the repo.

Patch-set: 1