Update patch set 1
Patch Set 1: -Code-Review

Kevin,

If you are going to be rude, please also try to be correct.

It is built because it is specified in the BoardCommonConfig in msm8974-common.

It builds correctly for me because I apparently did fix it when I first looked at the bringup and had forgotten.

Being polite and quoting a specific error would be more productive.

Patch-set: 1