Update patch set 1
Patch Set 1:

This is fine... mondrianwifi never even made it to N. If it ever makes it to O we can fragment makefiles that the subtrees can include into wifi/{bcmdhd,prima}/ (like we have already done for fingerprint/radio on klte). If power stats are even forward ported.

Now we just wait for rashed to fix his code.

Patch-set: 1
Label: Code-Review=0