Update patch set 4
Patch Set 4:

I know that sepolicy requires it but I need to be able to override WITH_DEXPREOPT because my build environment hangs because of dex2oatd at the end of my personal builds. When I run breakfast klte && make bacon WITH_DEXPREOPT=false it gets overriden and I think it is because of the removal of the ifeq.
Thoughts/suggestions?

Patch-set: 4