Update patch set 1
Patch Set 1:

Without this, Mobile data is not working on some devices.
The mobile connection icon shows an x, and there is little or no connectivity of mobile internet. I can confirm that this worked on the GT-I9506.
This was the case of other msm89xx common repositories as well, and some of them have allready added this to prevent this mobile data issue.
Have a look at the comments in this commit for instance:
https://review.lineageos.org/#/c/LineageOS/android_device_bq_msm8916-common/+/168163/

Patch-set: 1
Reviewer: Kay Arnesen <15397@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>